+++ frontend/assets/js/core/constants.js
@@ -1,1 +1,11 @@
-window.SmartRisk=window.SmartRisk||{};SmartRisk.Constants={APP_NAME:"SmartRisk CZ5",APP_VERSION:"2.0.0-rc1",STORAGE_PREFIX:"smartrisk.cz5"};
+(function () {
+  "use strict";
+
+  window.SmartRisk = window.SmartRisk || {};
+
+  window.SmartRisk.Constants = Object.freeze({
+    APP_NAME: "SmartRisk CZ5",
+    APP_VERSION: "2.1.0-alpha",
+    STORAGE_PREFIX: "smartrisk.cz5"
+  });
+})();