+++ frontend/assets/js/core/utils.js
@@ -1,1 +1,34 @@
-window.SmartRisk=window.SmartRisk||{};SmartRisk.Utils={formatDate(){return new Intl.DateTimeFormat("es-EC",{dateStyle:"long"}).format(new Date())},setTitle(t){document.title=t+" | SmartRisk CZ5"}};
+(function () {
+  "use strict";
+
+  window.SmartRisk = window.SmartRisk || {};
+
+  window.SmartRisk.Utils = {
+    formatDate(date = new Date()) {
+      return new Intl.DateTimeFormat("es-EC", {
+        dateStyle: "long"
+      }).format(date);
+    },
+
+    setTitle(title) {
+      document.title = `${title} | ${window.SmartRisk.Constants.APP_NAME}`;
+    },
+
+    normalizeText(value) {
+      return String(value ?? "")
+        .normalize("NFD")
+        .replace(/[\u0300-\u036f]/g, "")
+        .toLowerCase()
+        .trim();
+    },
+
+    escapeHtml(value) {
+      return String(value)
+        .replaceAll("&", "&amp;")
+        .replaceAll("<", "&lt;")
+        .replaceAll(">", "&gt;")
+        .replaceAll('"', "&quot;")
+        .replaceAll("'", "&#039;");
+    }
+  };
+})();